feat: Enable GQL Codegen Logic
