+++ .gitignore
@@ -11,6 +11,7 @@ node_modules
 dist
 dist-ssr
 *.local
+*.generated.ts
 
 # Editor directories and files
 .vscode/*
