+++ package.json
@@ -6,7 +6,8 @@
   "scripts": {
     "dev": "vite",
     "build": "vue-tsc && vite build",
-    "preview": "vite preview"
+    "preview": "vite preview",
+    "codegen": "graphql-codegen --config codegen.yml"
   },
   "dependencies": {
     "@apollo/client": "^3.7.12",
@@ -17,6 +18,12 @@
     "vue-router": "4"
   },
   "devDependencies": {
+    "@graphql-codegen/cli": "^4.0.1",
+    "@graphql-codegen/near-operation-file-preset": "^2.5.0",
+    "@graphql-codegen/typescript": "^4.0.0",
+    "@graphql-codegen/typescript-operations": "^4.0.0",
+    "@graphql-codegen/typescript-vue-apollo": "^3.3.7",
+    "@graphql-tools/schema": "^10.0.0",
     "@types/node": "^18.15.13",
     "@vitejs/plugin-vue": "^4.1.0",
     "typescript": "^5.0.2",
